Complete rework
